update code with BC break psr for compatiblity with symfony 3
a begin draft for maintainability with symfony 3 and change
autoloading of project for psr4
