+++ .travis.yml
@@ -1,11 +1,8 @@
 language: php
 
 php:
-    - 5.3
-    - 5.4
-    - 5.5
-    - 5.6
     - 7.0
+    - 7.1
 
 
 cache:
@@ -14,15 +11,9 @@ cache:
 
 matrix:
     include:
-        - php: 5.3
-          env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
-        - php: 5.4
-          env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
-        - php: 5.5
-          env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
-        - php: 5.6
-          env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
         - php: 7.0
+          env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
+        - php: 7.1
           env: 'COMPOSER_FLAGS="--prefer-lowest --prefer-stable"'
 
 branches:
