+++ composer.json
@@ -16,13 +16,13 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.3",
-        "symfony/http-kernel": "~2.5",
-        "symfony/dependency-injection": "~2.5",
-        "symfony/http-foundation": "~2.5",
-        "symfony/config": "~2.5",
-        "symfony/form": "~2.5",
-        "symfony/validator": "~2.5",
+        "php": ">=7.0",
+        "symfony/http-kernel": "~3.3",
+        "symfony/dependency-injection": "~3.3",
+        "symfony/http-foundation": "~3.3",
+        "symfony/config": "~3.3",
+        "symfony/form": "~3.3",
+        "symfony/validator": "~3.3",
         "twig/twig": "^1.12"
     },
     "require-dev": {
@@ -34,7 +34,12 @@
         "bin-dir": "bin"
     },
     "autoload": {
-        "psr-0": { "Knp\\DictionaryBundle": "src/" }
+        "psr-4": { "Knp\\DictionaryBundle\\": "./src" }
+    },
+    "autoload-dev": {
+        "psr-0": {
+            "spec": "./spec"
+        }
     },
     "extra": {
         "branch-alias": {
