+++ spec/Knp/DictionaryBundle/Dictionary/DictionarySpec.php
@@ -8,32 +8,33 @@ class DictionarySpec extends ObjectBehavior
 {
     public function let()
     {
-        $this->beConstructedWith('foo', array(
+        $this->beConstructedWith('foo', [
             'foo' => 0,
             'bar' => 1,
             'baz' => 2,
-        ));
+        ]);
     }
-
+    
     public function it_is_initializable()
     {
         $this->shouldHaveType('Knp\DictionaryBundle\Dictionary\Dictionary');
     }
-
+    
     public function its_getvalues_should_return_dictionary_values()
     {
-        $this->getValues()->shouldReturn(array(
+        $this->getValues()->shouldReturn([
             'foo' => 0,
             'bar' => 1,
             'baz' => 2,
-        ));
+        ])
+        ;
     }
-
+    
     public function its_getname_should_return_dictionary_name()
     {
         $this->getName()->shouldReturn('foo');
     }
-
+    
     public function it_access_to_value_like_an_array()
     {
         expect($this['foo']->getWrappedObject())->toBe(0);
