+++ spec/Knp/DictionaryBundle/Form/Type/DictionaryTypeSpec.php
@@ -2,8 +2,8 @@
 
 namespace spec\Knp\DictionaryBundle\Form\Type;
 
-use PhpSpec\ObjectBehavior;
 use Knp\DictionaryBundle\Dictionary\DictionaryRegistry;
+use PhpSpec\ObjectBehavior;
 
 class DictionaryTypeSpec extends ObjectBehavior
 {
@@ -11,7 +11,7 @@ class DictionaryTypeSpec extends ObjectBehavior
     {
         $this->beConstructedWith($registry);
     }
-
+    
     public function it_is_initializable()
     {
         $this->shouldHaveType('Knp\DictionaryBundle\Form\Type\DictionaryType');
