+++ spec/Knp/DictionaryBundle/Twig/DictionaryExtensionSpec.php
@@ -8,28 +8,29 @@ use PhpSpec\ObjectBehavior;
 
 class DictionaryExtensionSpec extends ObjectBehavior
 {
-    public function let(DictionaryRegistry $registry, Dictionary $dico1, Dictionary $dico2)
-    {
+    public function let(
+        DictionaryRegistry $registry,
+        Dictionary $dico1,
+        Dictionary $dico2
+    ) {
         $this->beConstructedWith($registry);
-
         $registry->get('test')->willReturn($dico1);
         $registry->get('other')->willReturn($dico2);
-
         $dico1->offsetGet('foo')->willReturn('bar');
         $dico2->offsetGet('foo')->willReturn(false);
     }
-
+    
     public function it_is_initializable()
     {
         $this->shouldHaveType('Knp\DictionaryBundle\Twig\DictionaryExtension');
     }
-
+    
     public function it_returns_a_dictionary_by_its_name($dico1, $dico2)
     {
         $this->getDictionary('test')->shouldReturn($dico1);
         $this->getDictionary('other')->shouldReturn($dico2);
     }
-
+    
     public function it_returns_a_value_from_a_dictionary()
     {
         $this->getValue('foo', 'test')->shouldReturn('bar');
