+++ src/Exception/DictionaryNotFoundException.php
