+++ src/Form/Type/DictionaryType.php
@@ -1,0 +1,54 @@
+<?php
+
+namespace Knp\DictionaryBundle\Form\Type;
+
+use Knp\DictionaryBundle\Dictionary\DictionaryRegistry;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\OptionsResolver\Options;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class DictionaryType extends AbstractType
+{
+    /**
+     * @var DictionaryRegistry
+     */
+    private $registry;
+    
+    /**
+     * @param DictionaryRegistry $registry
+     */
+    public function __construct(DictionaryRegistry $registry)
+    {
+        $this->registry = $registry;
+    }
+    
+    /**
+     * @param OptionsResolver $resolver
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $choices = function (Options $options) {
+            $name = $options['name'];
+            
+            return $this->registry[$name]->getValues();
+        };
+        $resolver->setDefaults(['name' => null, 'choices' => $choices]);
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function getParent()
+    {
+        return ChoiceType::class;
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'dictionary';
+    }
+}
