--- src/Knp/DictionaryBundle/DependencyInjection/KnpDictionaryExtension.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Knp\DictionaryBundle\DependencyInjection;
-
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-
-class KnpDictionaryExtension extends Extension
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function load(array $config, ContainerBuilder $container)
-    {
-        $configuration = new Configuration();
-        $config        = $this->processConfiguration($configuration, $config);
-        $container->setParameter('knp_dictionary.configuration', $config);
-
-        $loader = new YamlFileLoader(
-            $container,
-            new FileLocator(__DIR__ . '/../Resources/config')
-        );
-        $loader->load('dictionary.yml');
-    }
-}
