--- src/Knp/DictionaryBundle/Dictionary/ValueTransformer/TransformerInterface.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Knp\DictionaryBundle\Dictionary\ValueTransformer;
-
-interface TransformerInterface
-{
-    /**
-     * @param mixed $value
-     *
-     * @return bool
-     */
-    public function supports($value);
-
-    /**
-     * @param mixed $value
-     *
-     * @return mixed
-     */
-    public function transform($value);
-}
