--- src/Knp/DictionaryBundle/Resources/views/layout.html.twig
@@ -1,33 +1,0 @@
-{% extends 'WebProfilerBundle:Profiler:layout.html.twig' %}
-
-{% block menu %}
-<span class="label">
-<span class="icon">
-    <img width="26" height="28" alt="Dictionary" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABwAAAAcCAMAAABF0y+mAAAByFBMVEUAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACwtwJvAAAAAXRSTlMAQObYZgAAAAFiS0dEAIgFHUgAAAAJcEhZcwAAI28AACNvATX8B/sAAAAHdElNRQffAxAPNTnObSSWAAAANklEQVQoz2NgoCmYPB0FxCBJTccAqJKJ+CQZRiVpJjmdXEk86QBDrhGfJLLAtOl4jcLjIKoDACWgunBRwjJ0AAAAAElFTkSuQmCC">
-</span>
-<strong>Dictionaries</strong>
-<span class="count">
-    <span>{{ collector.dictionaries|length }}</span>
-</span>
-{% endblock %}
-
-{% block panel %}
-<h2>Dictionaries</h2>
-{% for dictionary in collector.dictionaries %}
-    <h3>{{ dictionary.name }}</h3>
-    <table>
-        <tbody>
-            <tr>
-                <th>Key</th>
-                <th>Value</th>
-            </tr>
-            {% for key, value in dictionary %}
-            <tr>
-                <td>{{ key }}</td>
-                <td>{{ value }}</td>
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-{% endfor %}
-{% endblock %}
