+++ src/KnpDictionaryBundle.php
