+++ src/Resources/config/dictionary.yml
@@ -1,0 +1,59 @@
+parameters:
+    knp_dictionary.data_collector.dictionary_data_collector.class:           Knp\DictionaryBundle\DataCollector\DictionaryDataCollector
+    knp_dictionary.data_collector.dictionary_data_collector.template:        KnpDictionaryBundle::layout.html.twig
+    knp_dictionary.dictionary.dictionary.class:                              Knp\DictionaryBundle\Dictionary\Dictionary
+    knp_dictionary.dictionary.dictionary_registry.class:                     Knp\DictionaryBundle\Dictionary\DictionaryRegistry
+    knp_dictionary.dictionary.value_transformer.constant_transformer.class:  Knp\DictionaryBundle\Dictionary\ValueTransformer\ConstantTransformer
+    knp_dictionary.form.type.dictionary_type.class:                          Knp\DictionaryBundle\Form\Type\DictionaryType
+    knp_dictionary.validator.constraints.dictionary_validator.class:         Knp\DictionaryBundle\Validator\Constraints\DictionaryValidator
+    knp_dictionary.twig.dictionary_extension.class:                          Knp\DictionaryBundle\Twig\DictionaryExtension
+    knp_dictionary.data_fixtures.dictionary_provider.class:                  Knp\DictionaryBundle\DataFixtures\DictionaryProvider
+services:
+    knp_dictionary.data_collector.dictionary_data_collector:
+        class: %knp_dictionary.data_collector.dictionary_data_collector.class%
+        arguments:
+            - '@knp_dictionary.registry'
+        tags:
+            - { name: data_collector, template: %knp_dictionary.data_collector.dictionary_data_collector.template%, id: 'dictionary' }
+
+    knp_dictionary.dictionary.dictionary_factory:
+        class: Knp\DictionaryBundle\Dictionary\DictionaryFactory
+
+    knp_dictionary.dictionary.dictionary_registry:
+        class: %knp_dictionary.dictionary.dictionary_registry.class%
+
+    knp_dictionary.dictionary.value_transformer.constant_transformer:
+        class: %knp_dictionary.dictionary.value_transformer.constant_transformer.class%
+        tags:
+            - { name: knp_dictionary.value_transformer }
+
+    knp_dictionary.form.type.dictionary_type:
+        class: %knp_dictionary.form.type.dictionary_type.class%
+        arguments:
+            - '@knp_dictionary.registry'
+        tags:
+            - { name: form.type, alias: dictionary }
+
+    knp_dictionary.registry:
+        alias: knp_dictionary.dictionary.dictionary_registry
+
+    knp_dictionary.validator.constraints.dictionary_validator:
+        class: %knp_dictionary.validator.constraints.dictionary_validator.class%
+        arguments:
+            - '@knp_dictionary.registry'
+        tags:
+            - { name: validator.constraint_validator, alias: knp_dictionary.dictionary_validator }
+
+    knp_dictionary.twig.dictionary_extension:
+        class: %knp_dictionary.twig.dictionary_extension.class%
+        arguments:
+            - '@knp_dictionary.registry'
+        tags:
+            - { name: twig.extension }
+
+    knp_dictionary.data_fixtures.dictionary_provider:
+        class: %knp_dictionary.data_fixtures.dictionary_provider.class%
+        arguments:
+            - '@knp_dictionary.registry'
+        tags:
+            - { name: knp_rad_fixtures_load.provider }
